piedpiper-ansible-runner-function/requirements.txt
git+git://github.com/ragingpastry/ansible-runner.git@fix-extra-vars
ansible
PyYAML
openshift
gunicorn
